Fix : - Change Print code button tooltip - Move fetcher in a PrintCodeFetcher file
